linting- type changes
